Refactorización de la lógica de check-in: eliminación de códigos de error sin uso y optimización del manejo de marcas temporales de los eventos.
